Add formal Lean specs and proofs for four constants. (#93)
Added an appropriate mathematical Lean constant d to Defs.lean, which represents the numerical value of the key parameter in the defining equation of the twisted Edwards curve:

− x 2 + y 2 = 1 + d x 2 y 2 .

Making use of the introduced Lean constant d , I formulated and proved formal Lean specs for:

    Constants/EDWARDS_D.lean,
    Constants/EDWARDS_D2.lean,
    Constants/EDWARDS_D_MINUS_ONE_SQUARED.lean,
    Constants/ONE_MINUS_EDWARDS_D_SQUARED.lean.

In the case of ONE_MINUS_EDWARDS_D_SQUARED.lean, I believe to have detected an error in the explanatory comment associated with this constant in the original Rust code, as the comment there appears to suggest that the constant is congruent to ( 1 − d ) 2 , when in reality I believe this should be 1 − d ² instead. I am mentioning this potential Rust-documentation-error in a brief note in ONE_MINUS_EDWARDS_D_SQUARED.lean.