scripts/extract-lean.sh: Fix sed string replacement / tweaks.
Previously, the script failed to correctly apply the substitutions via
bash string replacement (on MacOS, at least).
This commit instead uses sed for both the regex and non-regex case,
taking care to first properly escape the pattern string to prevent
inadvertent use of regex metacharacters / patterns.
This correction is implemented in the quoteRe() function.

Additionally, quoteSubst() corrects strings that are intended to be
used for replacement to escape special characters related to
referencing the regular expression pattern (&, match group indices).
This is applied unconditionally.

This ultimately shortens the code to apply tweaks, but adds two new
functions to the script.