.typos.toml: Don't check any scripts for typos.
This check was inherited from CI for Rust code, and should probably
not be applied to anything but the Rust parts of the project anyway.

This fixes a particular instance of a 'typo' that arises in some
scripts involving sed, where options/patterns may result in a false
positive 'typo' error.